pyproject.toml
[project]
name = "fednestd"
version = "0.1.0"
description = "Federated nested MoE framework (Tier1/2, dual governance, Kafka, etc.)"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Your Name", email = "you@example.com" }
]
dependencies = [
  "typer[all]>=0.12.0",
  "pydantic>=2.0",
  "torch>=2.0",
  "deepspeed>=0.14.0",
  "kafka-python>=2.0.2",
  "prometheus-client>=0.15.0",
  "rich>=13.0.0",
  "pydantic-settings>=2.0.0",  # Good for CLI config management
  # ... your other deps (flower, mlflow, datahub, ranger client, etc.)
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"] # pycodestyle, pyflakes, isort, pyupgrade

[tool.mypy]
strict = true
ignore_missing_imports = true

[build-system]
requires = ["hatchling"]  # or "poetry-core"
build-backend = "hatchling.build"

# CLI entrypoint: `fednestd` in shell
[project.scripts]
fednestd = "fednestd.__main__:main"